requirements.txt
fastapi
uvicorn
python-dotenv
click
cryptography
requests
google-genai

